backend/requirements.txt
tensorflow
fastapi
uvicorn[standard]
pillow
numpy
python-multipart
aiofiles

